requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
torch>=2.6.0
torchvision>=0.21.0

Pillow>=10.0.0
numpy>=1.24.0
python-dotenv==1.0.0
Werkzeug==2.3.7